refactor: 调整包路径并创建auth模块
