+++ app/api/route/auth/module.go
@@ -1,0 +1,1 @@
+package auth
